wcte12xp: Set the syncsrc in the span appropriately.
Before this commit, dahdi_tool would report "Internally clocked" for boards
supported by the wcte12xp driver both when receiving timing from the span and
providing it to the span.  Now it reports "Internally clocked" if providing
timeing to the span, and the card if receiving timeing from the span.
DAHDI-65.