wcfxo: Reset the DAA on module initialization.
The X100p and clones will sometimes work and sometimes not depending on wether
the DAA powers up in running state- this seems to be related to the power
supply. This problem is caused by the driver not reseting the DAA and may be
the source of a great many intermittent problems with this card.

(closes issue #14232)
Reported by: tallen8840
Patch by: tallen8840
Tested by: explidous, Flavio