Don't dereference the span after a channel was closed.
Releasing the channel may trigger releasing a span in certain cases
(e.g. the last channel on a xpp (Astribank). This gives a niec panic on
closing of Asterisk under the right flow.

As the span was only dereferenced to access the owner field, we can save
it locally.

Issue has been present since r6794 .